requirements-test.txt
pandas~=2.2.2
pytest~=8.3.3
apache-airflow[amazon,postgres]~=2.10